Revert "Split maze Subtypes into Subtypes and ResultsHandlers"
