+++ apps/src/maze/wordsearch.js
@@ -14,10 +14,6 @@ export default class WordSearch extends Subtype {
     this.map_ = config.level.map;
   }
 
-  getVisited() {
-    return this.visited_;
-  }
-
   /**
    * @override
    */
@@ -30,6 +26,21 @@ export default class WordSearch extends Subtype {
    */
   createDrawer(svg) {
     this.drawer = new WordSearchDrawer(this.maze_.map, '', svg);
+  }
+
+  /**
+   * Returns true if we've spelled the right word.
+   * @override
+   */
+  succeeded() {
+    return this.visited_ === this.goal_;
+  }
+
+  /**
+   * @override
+   */
+  shouldCheckSuccessOnMove() {
+    return false;
   }
 
   /**
